Revert "moved Name to api-common, alongside dependencies"
This reverts commit 3d66ce3ca199296fc01164d38a42f0ab65ca643f.
