+++ crates/cloud/src/galleries/routes.rs
@@ -1,7 +1,7 @@
-use crate::api::Name;
 use crate::app_data::AppData;
 use crate::auth;
 use crate::errors::UserError;
+use crate::utils::Name;
 use actix_web::{delete, get, patch, post, HttpRequest};
 use actix_web::{web, HttpResponse};
