+++ crates/cloud/src/utils.rs
@@ -29,6 +29,72 @@ use crate::{
     network::topology::{self, TopologyActor},
 };
 
+//TODO: Move this somewhere more appropriate
+use serde::de::{self, Visitor};
+use serde::{Deserialize, Deserializer, Serialize};
+use std::fmt;
+
+#[derive(Serialize, Clone, Debug)]
+pub struct Name(String);
+
+impl Name {
+    pub fn new(name: &str) -> Self {
+        Self(String::from(name))
+    }
+
+    pub fn as_str(&self) -> &str {
+        &self.0
+    }
+}
+
+struct NameVisitor;
+
+impl<'de> Visitor<'de> for NameVisitor {
+    type Value = Name;
+
+    fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
+        formatter
+            .write_str("a string that matches the regex: ^[\\w\\d_][\\w\\d_ \\(\\)\\.,'\\-!]*$")
+    }
+
+    fn visit_str<E>(self, value: &str) -> Result<Self::Value, E>
+    where
+        E: de::Error,
+    {
+        let max_len = 50;
+        let min_len = 1;
+        let char_count = value.chars().count();
+
+        lazy_static! {
+            static ref NAME_REGEX: Regex = Regex::new(r"^[\w\d_][\w\d_ \(\)\.,'\-!]*$").unwrap();
+        }
+
+        if char_count < min_len || char_count > max_len {
+            return Err(E::custom(format!(
+                "Name must be between {} and {} characters",
+                min_len, max_len
+            )));
+        }
+        if !NAME_REGEX.is_match(value) {
+            return Err(E::custom("Name contains invalid characters"));
+        }
+
+        if value.is_inappropriate() {
+            return Err(E::custom("Name contains inappropriate content"));
+        }
+
+        Ok(Name(value.to_owned()))
+    }
+}
+impl<'de> Deserialize<'de> for Name {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        deserializer.deserialize_string(NameVisitor)
+    }
+}
+
 pub(crate) fn on_room_changed(
     network: &Addr<TopologyActor>,
     cache: &Arc<RwLock<LruCache<api::ProjectId, ProjectMetadata>>>,
